+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -182,7 +182,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KWGTD7RX8DF4E9CJ7Y5K13TR"),
+  _id: z.string().default("01KWH6BZ6SGPBW2J21RJQF57WE"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: ListDatasourcesStatus$inboundSchema,