+++ packages/orq-rc/RELEASES.md
@@ -10469,3 +10469,13 @@ Based on:
 - [typescript v4.11.0-rc.69] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.69] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.69 - packages/orq-rc
+
+## 2026-07-02 10:38:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.787.0 (2.914.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.70] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.70] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.70 - packages/orq-rc